rename task + CLAUDE.md
