+++ loading_pipeline/lib/tasks/exports/write_new_variants_parquet.py
@@ -19,7 +19,7 @@ from loading_pipeline.lib.tasks.exports.misc import (
     unmap_formatting_annotation_enums,
 )
 from loading_pipeline.lib.tasks.files import GCSorLocalTarget
-from loading_pipeline.lib.tasks.update_variant_annotations_table_with_new_samples import (
+from loading_pipeline.lib.tasks.update_variant_annotations_table_with_new_variants import (
     UpdateVariantAnnotationsTableWithNewVariantsTask,
 )
 from loading_pipeline.lib.tasks.write_new_variants_table import (
