+++ loading_pipeline/lib/tasks/update_variant_annotations_table_with_new_variants.py
